Add full support for hosted docs in sync and async modes Standardize on a universal get_doc which applies to any hosted or async docs
Why is this change needed?
--------------------------
With hosted docs coming and the limitations of OpenAPI/Swagger 2, we need to add additional methods to handle the different endpoints and payloads required.

How does it address the issue?
------------------------------
Fills out the methods required to support all the combinations.  We now have:

- create_doc (returns binary)
- create_hosted_doc (returns payload with download_id)
- create_async_doc (returns payload with status_id)
- get_async_doc_status (uses status id to get status with download_id if done)
- get_doc (uses download_id to retrieve binary doc)

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------